policy: remove runtime governance matrix crate
