+++ .github/workflows/microcrate-ci.yml
@@ -149,7 +149,6 @@ jobs:
             adze-governance-runtime-core
             adze-governance-runtime-reporting
             adze-feature-policy-core
-            adze-runtime-governance-matrix
             adze-bdd-governance-core
             adze-bdd-grammar-fixtures
             adze-bdd-grid-core
@@ -224,7 +223,6 @@ jobs:
             -p adze-governance-runtime-core \
             -p adze-governance-runtime-reporting \
             -p adze-feature-policy-core \
-            -p adze-runtime-governance-matrix \
             -- --test-threads=$RUST_TEST_THREADS
 
   test-bdd:
