--- crates/runtime-governance-matrix/README.md
@@ -1,10 +1,0 @@
-# adze-runtime-governance-matrix
-
-Runtime-capability governance matrix with feature-flag profiles and BDD reporting.
-
-Part of the [adze](https://github.com/EffortlessMetrics/adze) workspace.
-
-## License
-
-Licensed under either of [Apache License, Version 2.0](../../LICENSE-APACHE)
-or [MIT License](../../LICENSE-MIT) at your option.
