+++ runtime2/Cargo.toml
@@ -16,7 +16,7 @@ thiserror = "1.0"
 smallvec.workspace = true
 rustc-hash.workspace = true
 adze-parsetable-metadata = { version = "0.1.0", path = "../crates/parsetable-metadata" }
-adze-runtime-governance-matrix = { version = "0.1.0", path = "../crates/runtime-governance-matrix" }
+adze-governance-runtime-core = { version = "0.1.0", path = "../crates/governance-runtime-core" }
 adze-linecol-core = { version = "0.1.0", path = "../crates/linecol-core" }
 
 # GLR core (optional, enabled by default)
